backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
google-genai
python-dotenv
websockets
